# Use bytes package to improve performance