It takes 36% less memory, and 9.4% less allocations.

```
$ benchstat old.txt new.txt
name          old time/op    new time/op    delta
UserAgent-12     217µs ± 2%     216µs ± 2%     ~     (p=0.315 n=10+10)

name          old alloc/op   new alloc/op   delta
UserAgent-12    42.5kB ± 0%    27.1kB ± 0%  -36.11%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
UserAgent-12     1.52k ± 0%     1.37k ± 0%   -9.43%  (p=0.000 n=10+10)
```

<details>

<summary>old.txt</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/mileusna/useragent
cpu: Intel(R) Core(TM) i5-10600 CPU @ 3.30GHz
BenchmarkUserAgent-12    	    5184	    214837 ns/op	   42490 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5476	    214017 ns/op	   42445 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5276	    214757 ns/op	   42505 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5454	    218415 ns/op	   42433 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5116	    217853 ns/op	   42408 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5482	    220742 ns/op	   42479 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5217	    218494 ns/op	   42431 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5294	    216651 ns/op	   42413 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5479	    218571 ns/op	   42472 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5474	    216790 ns/op	   42452 B/op	    1517 allocs/op
PASS
ok  	github.com/mileusna/useragent	12.049s
```

</details>

<details>

<summary>new.txt</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/mileusna/useragent
cpu: Intel(R) Core(TM) i5-10600 CPU @ 3.30GHz
BenchmarkUserAgent-12    	    5372	    216292 ns/op	   27116 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5419	    214496 ns/op	   27101 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5529	    216486 ns/op	   27091 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5617	    214828 ns/op	   27108 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5670	    214217 ns/op	   27100 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5637	    214188 ns/op	   27134 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5559	    213480 ns/op	   27143 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5548	    221259 ns/op	   27164 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5287	    219650 ns/op	   27134 B/op	    1374 allocs/op
BenchmarkUserAgent-12    	    5336	    218161 ns/op	   27132 B/op	    1374 allocs/op
PASS
ok  	github.com/mileusna/useragent	12.343s
```

</details>